feat/test serializer

+++ spec/poros/current_weather_spec.rb
@@ -11,9 +11,9 @@ RSpec.describe CurrentWeather, :vcr do
     expect(object.sunset).to be_a String
     expect(object.temperature).to be_a Float
     expect(object.feels_like).to be_a Float
-    expect(object.humidity).to be_a Integer
-    expect(object.uvi).to be_a Integer
-    expect(object.visibility).to be_a Integer
+    expect(object.humidity).to be_a Numeric
+    expect(object.uvi).to be_a Numeric
+    expect(object.visibility).to be_a Numeric
     expect(object.conditions).to be_a String
     expect(object.icon).to be_a String
   end
